Add files via upload
The Temperature is the one which is mainly considered for the changing weather conditions. This paper deals
with the analysis of Temperature, where we use Big Data features like MapReduce to calculate the Average
Temperature with respect to its year. We use HIVE, which is an open-source data warehousing software used for
reading large datasets. Using HIVE, we write queries to fetch the data and we can also update it. Further, we use
Random Forest Regression Technique, a Machine Learning Algorithm to classify and predict the temperature
based on the classification. The dataset for this analysis is taken from the Data.gov website which contained data
for about 85 years and we have taken the data for about 30 years from 1990 – 2020. The above analysis is done
using Hadoop and Pyspark